backend-plugin-api: internal refactor to split creators
