+++ packages/backend-plugin-api/src/wiring/createExtensionPoint.test.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2022 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { createExtensionPoint } from './createExtensionPoint';
+
+describe('createExtensionPoint', () => {
+  it('should create an ExtensionPoint', () => {
+    const extensionPoint = createExtensionPoint({ id: 'x' });
+    expect(extensionPoint).toBeDefined();
+    expect(extensionPoint.id).toBe('x');
+    expect(() => extensionPoint.T).not.toThrow();
+    expect(String(extensionPoint)).toBe('extensionPoint{x}');
+  });
+});
